+++ kissom_pg/pgConstants.py
@@ -1,0 +1,4 @@
+PG_TABLE = "pg_tables"
+PG_VIEW = "pg_views"
+PG_MATERIALIZED_VIEW = "pg_matviews"
+PG_UNKNOWN = "unknown_catalog_type"